Break up the generator to be more readable
