+++ Sources/SwiftBuffet/Documentation.docc/GettingStarted.md
@@ -16,7 +16,7 @@ targets: [
     .target(
         name: "YourTarget",
         resources: [
-            .process("models.proto")
+            .process("buffet.proto")
         ],
         plugins: [
             .plugin(name: "SwiftBuffetPlugin", package: "SwiftBuffet")
@@ -30,23 +30,23 @@ That's it. On every build, the plugin finds each `.proto` file in the target, ru
 ```proto
 syntax = "proto3";
 
-message Person {
-    string name = 1;
-    int32 id = 2;
-    string email = 3;
+message FingerSandwich {
+    string filling = 1;
+    string bread = 2;
+    optional int32 quarters = 3;
 }
 
-message AddressBook {
-    repeated Person people = 1;
-    optional bool is_current = 2;
+message Buffet {
+    repeated FingerSandwich sandwiches = 1;
+    optional bool crusts_removed = 2;
 }
 ```
 
 ## Use the generated types
 The example above generates two plain structs you can use anywhere in your target:
 ```swift
-let person = Person(name: "Robin", id: 7, email: "robin@example.com")
-let book = AddressBook(people: [person], isCurrent: true)
+let cucumber = FingerSandwich(filling: "cucumber", bread: "white", quarters: 4)
+let buffet = Buffet(sandwiches: [cucumber], crustsRemoved: true) // naturally
 ```
 Field names are converted from `snake_case` to `camelCase`, `repeated` fields become arrays, and proto3 `optional` fields become Swift optionals with sensible decoding defaults. See <doc:GeneratedCode> for the full mapping.
